dummy-kernel: Rename template to empty-kernel
Issue: LINCD-5300

Wind River is working to remove/replace non-inclusive language in our product.
For details on the InfoDev process for doing so, see:
https://confluence.wrs.com/pages/viewpage.action?spaceKey=InfoDev&title=Inclusive+Language+in+Technology

This defect addresses the use of "dummy", currently used for the "dummy kernel"
and linux-yocto-dummy recipe, which we recommend be changed to match our style
standards, "empty" or "test" kernel.

Signed-off-by: He Zhe <zhe.he@windriver.com>