feature/podman: remove ostree from whitelist
Issue: LINUXEXEC-16033 LINCD-6717

As the meta-virtualization maintains podman in development version,
so we maintain a latest release version whose recipe copied from the
upstream under meta-virtualization and podman depends on skopeo which
depends on ostree.

Considering we don't use ostree function currently and also not ready
to use the ostree under meta-oe, so remove the ostree currently.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>