systemd-conf: remove bbappend
Issue: LINCD-7702

If there are more than one NICs exist and one of them has no carrier,
the status of systemd-networkd-wait-online.service has been 'failed'. So
link 80-wired.network to /dev/null to solve that issue.

But it has changed the status of systemd-networkd-wait-online.service to
'activating' for this circumstance by systemd. So drop the local patch.

Signed-off-by: Kai Kang <kai.kang@windriver.com>