linux-libc-headers-custom: non-versioned branches are gone; point at v5.17
Issue: LINCD-8527

The linux-yocto-dev no longer uses/updates the non-versioned branches; they
are currently stuck on v5.16 and will be removed externally in the near
future.  We've already removed them internally in advance to catch issues
exactly like this.

Point at v5.17 -- we should automate this so we aren't having to update
on every linux-yocto-dev version change ~6 times a year.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>