linux-yocto: Set SRCREV to AUTOREV
Issue: LINCD-8891

After patches get merged in Yocto linux-yocto source tree, they still need
to wait for oe-core to bump up the SRCREV to take effect. This delay may be
two patch versions(as 33 in 5.15.33) or longer which is unnecessary for LINCD.
Plus, linux-yocto-rt for qemu BSPs and all hardware BSPs have already used the
latest source for quite a long time. Let's align with them.

Signed-off-by: He Zhe <zhe.he@windriver.com>