dhcpcd: allow more generic network interface specification
Issue: LIN1022-2727

We have a 'eth0' hardcode in dhcpcd. This is usually OK because
in WRLinux framework, the first NIC will be named eth0. However,
if a machine have multiple NICs and the first NIC is not connected,
then this hardcoding is going to cause some problem. For example,
eth2 is connected while eth0 is not. In such case, eth2 is not
configured.

To solve this issue, we will need a more generic network interface
specification method for dhcpcd. In /etc/default/dhcpcd, two variables
are allowed, INTERFACES and INTERFACE_REGEXES. INTERFACES is used to
list the interface names while INTERFACE_REGEXES is used to specify
the regexes to match interfaces. By default, we specify INTERFACE_REGEX
to be 'eth.* pfe.*', which matches all interfaces, and only the connected
ones will be configured. The INTERFACES variable is still allowed,
not only for backward compatibility but also help make life easier
for end users. For example, if the user only wants eth1 to be configured,
using INTERFACE_REGEX will not be easy to get right, because regex 'eth1'
will also match 'eth1.1'. So we choose to keep both vars, and by default
to use 'eth.* pfe.*' which matches all current known interfaces supported
by WRLinux. Note that the pfe.* comes from nxp-s32g.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>