wr-common-packages-native.inc: remove clang-native and p11-kit-native
Issue: LIN1022-1884

It is not clear explained why clang-native and p11-kit-native are added
to DEPENDS of wr-common-packages-native by commit:

http://lxgit.wrs.com/cgit/layers/wrlinux/commit/?h=WRLINUX_10_22_BASE&id=72fc647ab5bdc9a552f4ef335423db59cede3201

For p11-kit, there is no p11-kit-native.

And it fails to build wr-common-packages-native.inc with error:

ERROR: Nothing PROVIDES 'clang-native' (but
/path_to/wrlinux/recipes-base/wr-common-packages-native/wr-common-packages-native.bb
DEPENDS on or otherwise requires it)

So remove clang-native from wr-common-packages-native.inc too.

Signed-off-by: Kai Kang <kai.kang@windriver.com>