feature/kubernetes: remove hardcoded docker
Issue: LINUXEXEC-25377

Now that we've supported cri-o, we really don't need to put
docker as a dependency for this template. In addition, using
docker in this template puts extra unnecessary constraints.
For example, docker conflicts with podman, but we can of course
use k8s with cri-o and at the same time install podman.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>