wrlinux: move PREFERRED_VERSION from most templates
Now that WRL_RECIPES have been moved out of templates we have enabled
recipes being built without including templates. So we we can't rely
on PREFERRED_VERSION set in templates and need to set these in layer.conf
so they are always applied.

Some PREFERRED_VERSION remain in templates, but this is only when a
template exists to accomplish the goal of using a different version
than would typically be in use.

Issue: LINUXEXEC-25187

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>