dpdk: Add back to world build
Issue: LIN1022-3791

DPDK was excluded from world build since it depended on specific kernel
features and there was no way to check if the features have been included for
the build. See the following link for details.
http://lxgit.wrs.com/cgit/layers/wrlinux/commit/?h=master-wr&id=9a4b4800f57856485438d075740926bb24bbfc9e

Now DPDK was able to build without those kernel features, so it should be added
back to world build.

Signed-off-by: He Zhe <zhe.he@windriver.com>