gen-image: Cleanup old rpms in repos
The rpm repos' size increase quickly during each rcpl, remove the old ones when
there are multiple versions to reduce the size.

The packagefeed-stability.bbclass can remove old rpms, but it can only remove
old revisions, can't remove old versions, and it only removes the old revisions
packages when there are rpms added to the repos.

The cleanup-repos.py runs before the build in case we need support multiple
versions in the future, in that case the removed rpms would be built out again
after removed.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>